# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sales & Orders API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceDetail"

    delete:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel an order (only if not shipped)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── AR / Aging ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    InvoiceSummary:
      type: object
      required:
        [
          id,
          invoiceNumber,
          amount,
          amountPaid,
          balanceDue,
          invoiceDate,
          dueDate,
          isPaid,
        ]
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        amount:
          type: number
        amountPaid:
          type: number
        balanceDue:
          type: number
        invoiceDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        isPaid:
          type: boolean
        notes:
          type: ["string", "null"]
        orderId:
          type: ["integer", "null"]
    InvoiceParty:
      type: object
      required: [name, email, phone]
      properties:
        id:
          type: ["integer", "null"]
        name:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        repName:
          type: ["string", "null"]
    InvoiceOrder:
      type: object
      required:
        [
          id,
          orderNumber,
          orderDate,
          subtotal,
          discountTotal,
          shippingCost,
          total,
        ]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        orderDate:
          type: string
          format: date-time
        shippingMethod:
          type: ["string", "null"]
        trackingNumber:
          type: ["string", "null"]
        subtotal:
          type: number
        discountTotal:
          type: number
        shippingCost:
          type: number
        total:
          type: number
        customTerms:
          type: ["string", "null"]
    InvoiceDetail:
      allOf:
        - $ref: "#/components/schemas/InvoiceSummary"
        - type: object
          required: [customer, company, order, lineItems]
          properties:
            customer:
              $ref: "#/components/schemas/InvoiceParty"
            company:
              $ref: "#/components/schemas/InvoiceParty"
            order:
              oneOf:
                - $ref: "#/components/schemas/InvoiceOrder"
                - type: "null"
            lineItems:
              type: array
              items:
                $ref: "#/components/schemas/OrderLineItem"

    # ── Products ──────────────────────────────────────────────────────────────
...
    OrderLineItem:
      type: object
      required: [productId, sku, productName, quantity, unitPrice, lineTotal]
      properties:
        productId:
          type: integer
        sku:
          type: string
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        discountAmount:
          type: number
        lineTotal:
          type: number
        promotionName:
          type: ["string", "null"]
...
    Order:
      type: object
      required:
        [id, orderNumber, customerId, customerName, orderDate, status, total]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        repId:
          type: ["integer", "null"]
        repName:
          type: ["string", "null"]
        orderDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [open, in_transit, fulfilled, cancelled]
        total:
          type: number
        shippingPolicyId:
          type: ["integer", "null"]
        shippingCarrier:
          type: ["string", "null"]
        trackingNumber:
          type: ["string", "null"]
        shippingMethod:
          type: ["string", "null"]
        invoice:
          oneOf:
            - $ref: "#/components/schemas/InvoiceSummary"
            - type: "null"
...